refactor: renaming OakPupilJourneyPreviousLessons files
