+++ src/components/organisms/pupil/OakPupilJourneyListItemSubheading/index.ts
@@ -1,0 +1,1 @@
+export * from "./OakPupilJourneyListItemSubheading";
