--- src/components/organisms/pupil/OakPupilJourneyPreviousLessons/index.ts
@@ -1,1 +1,0 @@
-export * from "./OakPupilJourneyPreviousLessons";
